requirements.txt
gunicorn<21.0,>=19.2.0
flask<3.0,>=1.0
flask_restful==0.3.10
google-cloud-firestore==2.15.0
google-cloud-pubsub==2.19.4
wtforms==3.1.2
requests~=2.31.0
marshmallow==3.20.2